+++ src/index.js
@@ -1,6 +1,7 @@
 const Doz = require('./Doz');
 const collection = require('./collection');
 const component = require('./component');
+const use = require('./plugin').use;
 const Component = require('./component/Component').Component;
 const h = require('./vdom/h');
 const mixin = require('./component/global-mixin');
@@ -28,6 +29,10 @@ Object.defineProperties(Doz, {
     },
     mixin: {
         value: mixin,
+        enumerable: true
+    },
+    use: {
+        value: use,
         enumerable: true
     },
     version: {