+++ test/doz.on.js
@@ -1,7 +1,7 @@
 const Doz = require('../index');
 const be = require('bejs');
 
-describe('Doz.on.app.draw', function () {
+describe('Doz.on', function () {
 
     this.timeout(5000);
 
@@ -24,18 +24,28 @@ describe('Doz.on.app.draw', function () {
 
             document.body.innerHTML = `<div id="app"></div>`;
 
+            const myPluginAddCiao = function(Doz, app) {
+                Doz.mixin({
+                    myCiao() {
+                        return 'Ciao'
+                    }
+                });
+
+                app.on('appDraw', (next, prev) => {
+                    console.log(next);
+                    done();
+                })
+            };
+
+            Doz.use(myPluginAddCiao);
+
             Doz.component('salutation-card', {
                 template() {
                     return `
-                        <div>Hello ${this.props.title} ${this.props.name}</div>
+                        <div>${this.myCiao()} ${this.props.title} ${this.props.name}</div>
                     `
-                },
-                onAppDraw(next, prev) {
-                    console.log(next, prev);
-                    next.children[0] = 'Ciao'
                 }
             });
-
 
             new Doz({
                 root: '#app',
@@ -46,12 +56,6 @@ describe('Doz.on.app.draw', function () {
                     </salutation-card>
                 `
             });
-
-            setTimeout(function () {
-                console.log(document.body.innerHTML);
-                done();
-            }, 100);
-
         });
     });
 });